Refactor podfile analysis into it's own file Originally the podfile analysis was done in the 'cli_runner.rb' file.  The analyzer will be easier to modify with the code separated from the CLI runner options and BOM building process.
