+++ lib/cyclonedx/cocoapods/cli_runner.rb
@@ -18,19 +18,15 @@
 # Copyright (c) OWASP Foundation. All Rights Reserved.
 #
 
-require 'optparse'
 require 'logger'
-require 'cocoapods'
+require 'optparse'
 
-require_relative 'component'
-require_relative 'pod'
-require_relative 'pod_attributes'
-require_relative 'source'
 require_relative 'bom_builder'
+require_relative 'component'
+require_relative 'podfile_analyzer'
 
 module CycloneDX
   module CocoaPods
-    class PodfileParsingError < StandardError; end
     class BOMOutputError < StandardError; end
 
     class CLIRunner
@@ -41,10 +37,10 @@ module CycloneDX
           setup_logger(verbose: options[:verbose])
           @logger.debug "Running cyclonedx-cocoapods with options: #{options}"
 
-          podfile, lockfile = ensure_podfile_and_lock_are_present(options)
-          pods = parse_pods(podfile, lockfile)
-
-          populate_pods_with_additional_info(pods)
+          analyzer = PodfileAnalyzer.new(logger: @logger)
+          podfile, lockfile = analyzer.ensure_podfile_and_lock_are_present(options)
+          pods = analyzer.parse_pods(podfile, lockfile)
+          analyzer.populate_pods_with_additional_info(pods)
 
           bom = BOMBuilder.new(component: component_from_options(options), pods: pods).bom(version: options[:bom_version] || 1)
           write_bom_to_file(bom: bom, options: options)
@@ -56,6 +52,7 @@ module CycloneDX
 
 
       private
+
 
       def parseOptions
         parsedOptions = {}
@@ -115,90 +112,6 @@ module CycloneDX
       def setup_logger(verbose: true)
         @logger ||= Logger.new($stdout)
         @logger.level = verbose ? Logger::DEBUG : Logger::INFO
-      end
-
-
-      def ensure_podfile_and_lock_are_present(options)
-        project_dir = Pathname.new(options[:path] || Dir.pwd)
-        raise PodfileParsingError, "#{options[:path]} is not a valid directory." unless File.directory?(project_dir)
-        options[:podfile_path] = project_dir + 'Podfile'
-        raise PodfileParsingError, "Missing Podfile in #{project_dir}. Please use the --path option if not running from the CocoaPods project directory." unless File.exist?(options[:podfile_path])
-        options[:podfile_lock_path] = project_dir + 'Podfile.lock'
-        raise PodfileParsingError, "Missing Podfile.lock, please run 'pod install' before generating BOM" unless File.exist?(options[:podfile_lock_path])
-
-        initialize_cocoapods_config(project_dir)
-
-        lockfile = ::Pod::Lockfile.from_file(options[:podfile_lock_path])
-        verify_synced_sandbox(lockfile)
-
-        return ::Pod::Podfile.from_file(options[:podfile_path]), lockfile
-      end
-
-
-      def initialize_cocoapods_config(project_dir)
-        ::Pod::Config.instance.installation_root = project_dir
-      end
-
-
-      def verify_synced_sandbox(lockfile)
-        manifestFile = ::Pod::Config.instance.sandbox.manifest
-        raise PodfileParsingError, "Missing Manifest.lock, please run 'pod install' before generating BOM" if manifestFile.nil?
-        raise PodfileParsingError, "The sandbox is not in sync with the Podfile.lock. Run 'pod install' or update your CocoaPods installation." unless lockfile == manifestFile
-      end
-
-
-      def cocoapods_repository_source(podfile, lockfile, pod_name)
-        @source_manager ||= create_source_manager(podfile)
-        return Source::CocoaPodsRepository.searchable_source(url: lockfile.spec_repo(pod_name), source_manager: @source_manager)
-      end
-
-
-      def git_source(lockfile, pod_name)
-        checkout_options = lockfile.checkout_options_for_pod_named(pod_name)
-        url = checkout_options[:git]
-        [:tag, :branch, :commit].each do |type|
-          return Source::GitRepository.new(url: url, type: type, label: checkout_options[type]) if checkout_options[type]
-        end
-        return Source::GitRepository.new(url: url)
-      end
-
-
-      def source_for_pod(podfile, lockfile, pod_name)
-        root_name = pod_name.split('/').first
-        return cocoapods_repository_source(podfile, lockfile, root_name) unless lockfile.spec_repo(root_name).nil?
-        return git_source(lockfile, root_name) unless lockfile.checkout_options_for_pod_named(root_name).nil?
-        return Source::LocalPod.new(path: lockfile.to_hash['EXTERNAL SOURCES'][root_name][:path]) if lockfile.to_hash['EXTERNAL SOURCES'][root_name][:path]
-        return Source::Podspec.new(url: lockfile.to_hash['EXTERNAL SOURCES'][root_name][:podspec]) if lockfile.to_hash['EXTERNAL SOURCES'][root_name][:podspec]
-        return nil
-      end
-
-
-      def parse_pods(podfile, lockfile)
-        @logger.debug "Parsing pods from #{podfile.defined_in_file}"
-        return lockfile.pod_names.map do |name|
-          Pod.new(name: name, version: lockfile.version(name), source: source_for_pod(podfile, lockfile, name), checksum: lockfile.checksum(name))
-        end
-      end
-
-
-      def create_source_manager(podfile)
-        sourceManager = ::Pod::Source::Manager.new(::Pod::Config::instance.repos_dir)
-        @logger.debug "Parsing sources from #{podfile.defined_in_file}"
-        podfile.sources.each do |source|
-          @logger.debug "Ensuring #{source} is available for searches"
-          sourceManager.find_or_create_source_with_url(source)
-        end
-        @logger.debug "Source manager successfully created with all needed sources"
-        return sourceManager
-      end
-
-
-      def populate_pods_with_additional_info(pods)
-        pods.each do |pod|
-          @logger.debug "Completing information for #{pod.name}"
-          pod.complete_information_from_source
-        end
-        return pods
       end
